requirements.txt
# OmniParser dependencies
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
supervision>=0.17.0
huggingface-hub>=0.20.0

# For Florence-2 model
einops>=0.7.0
timm>=0.9.0
flash-attn>=2.3.0

# Web server
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0